# frontend/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BacktestBandRow": {
        "properties": {
          "band": {
            "$ref": "#/components/schemas/SignalBand"
          },
          "hit_rate": {
            "title": "Hit Rate",
            "type": "number"
          },
          "hits": {
            "title": "Hits",
            "type": "integer"
          },
          "max_drawdown": {
            "title": "Max Drawdown",
            "type": "number"
          },
          "misses": {
            "title": "Misses",
            "type": "integer"
          },
          "no_bets": {
            "title": "No Bets",
            "type": "integer"
          },
          "recommendations": {
            "title": "Recommendations",
            "type": "integer"
          },
          "roi": {
            "title": "Roi",
            "type": "number"
          },
          "units": {
            "title": "Units",
            "type": "number"
          }
        },
        "required": [
          "recommendations",
          "hits",
          "misses",
          "hit_rate",
          "units",
          "roi",
          "max_drawdown",
          "band",
          "no_bets"
        ],
        "title": "BacktestBandRow",
        "type": "object"
      },
      "BacktestReport": {
        "description": "Informe del backtest sobre historiales fuera de calibración.",
        "properties": {
          "by_band": {
            "items": {
              "$ref": "#/components/schemas/BacktestBandRow"
            },
            "title": "By Band",
            "type": "array"
          },
          "decisions": {
            "title": "Decisions",
            "type": "integer"
          },
          "house_edge_reference": {
            "title": "House Edge Reference",
            "type": "number"
          },
          "no_bet_rate": {
            "title": "No Bet Rate",
            "type": "number"
          },
          "no_bets": {
            "title": "No Bets",
            "type": "integer"
          },
          "overall": {
            "$ref": "#/components/schemas/BacktestTally"
          },
          "recommendations": {
            "title": "Recommendations",
            "type": "integer"
          },
          "sessions": {
            "title": "Sessions",
            "type": "integer"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "spins_evaluated": {
            "title": "Spins Evaluated",
            "type": "integer"
          },
          "threshold": {
            "title": "Threshold",
            "type": "number"
          },
          "weak_threshold": {
            "title": "Weak Threshold",
            "type": "number"
          }
        },
        "required": [
          "source",
          "sessions",
          "spins_evaluated",
          "decisions",
          "recommendations",
          "no_bets",
          "no_bet_rate",
          "threshold",
          "weak_threshold",
          "overall",
          "by_band",
          "house_edge_reference"
        ],
        "title": "BacktestReport",
        "type": "object"
      },
      "BacktestSource": {
        "description": "Sobre que historiales corre el backtest.\n\n`sessions` son datos reales, fuera de calibracion por construccion.\n`simulated` es la linea base: ruedas justas donde el ROI tiene que quedarse\nen la ventaja de la casa.",
        "enum": [
          "sessions",
          "simulated"
        ],
        "title": "BacktestSource",
        "type": "string"
      },
...
      "CreateGameRequest": {
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "type"
        ],
        "title": "CreateGameRequest",
        "type": "object"
      },
      "CreateGameVariantRequest": {
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "config": {
            "$ref": "#/components/schemas/GameVariantConfig-Input"
          },
          "house_edge": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "House Edge",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "house_edge",
          "config"
        ],
        "title": "CreateGameVariantRequest",
        "type": "object"
      },
...
      "GameResponse": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "variants": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/GameVariantResponse"
            },
            "title": "Variants",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "active",
          "variants"
        ],
        "title": "GameResponse",
        "type": "object"
      },
      "GameVariantConfig-Input": {
        "description": "El contenido completo de game_variants.categories_json + metadata.",
        "properties": {
          "allowed_combinations": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AllowedCombination"
            },
            "title": "Allowed Combinations",
            "type": "array"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/GameCategory-Input"
            },
            "minItems": 1,
            "title": "Categories",
            "type": "array"
          },
          "possible_outcomes": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Possible Outcomes",
            "type": "array"
          },
          "recommendation_threshold": {
            "default": 50,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Recommendation Threshold",
            "type": "number"
          },
          "weak_threshold": {
            "default": 35,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Weak Threshold",
            "type": "number"
          }
        },
        "required": [
          "possible_outcomes",
          "categories"
        ],
        "title": "GameVariantConfig",
        "type": "object"
      },
      "GameVariantConfig-Output": {
        "description": "El contenido completo de game_variants.categories_json + metadata.",
        "properties": {
          "allowed_combinations": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AllowedCombination"
            },
            "title": "Allowed Combinations",
            "type": "array"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/GameCategory-Output"
            },
            "minItems": 1,
            "title": "Categories",
            "type": "array"
          },
          "possible_outcomes": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Possible Outcomes",
            "type": "array"
          },
          "recommendation_threshold": {
            "default": 50,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Recommendation Threshold",
            "type": "number"
          },
          "weak_threshold": {
            "default": 35,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Weak Threshold",
            "type": "number"
          }
        },
        "required": [
          "possible_outcomes",
          "categories",
          "allowed_combinations",
          "recommendation_threshold",
          "weak_threshold"
        ],
        "title": "GameVariantConfig",
        "type": "object"
      },
      "GameVariantResponse": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "config": {
            "$ref": "#/components/schemas/GameVariantConfig-Output"
          },
          "game_id": {
            "format": "uuid",
            "title": "Game Id",
            "type": "string"
          },
          "house_edge": {
            "title": "House Edge",
            "type": "number"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "game_id",
          "name",
          "house_edge",
          "config",
          "active"
        ],
        "title": "GameVariantResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UpdateGameRequest": {
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "UpdateGameRequest",
        "type": "object"
      },
      "UpdateGameVariantRequest": {
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GameVariantConfig-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "house_edge": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "House Edge"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "UpdateGameVariantRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Sebasanalisis API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/games": {
      "post": {
        "operationId": "create_game_admin_games_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Game",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/games/{game_id}": {
      "patch": {
        "operationId": "update_game_admin_games__game_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Game Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Game",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/games/{game_id}/variants": {
      "post": {
        "operationId": "create_variant_admin_games__game_id__variants_post",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Game Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameVariantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameVariantResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Variant",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/games/{game_id}/variants/{variant_id}": {
      "patch": {
        "operationId": "update_variant_admin_games__game_id__variants__variant_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Game Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "variant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Variant Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGameVariantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameVariantResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Variant",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/admin/recommendations/backtest": {
      "get": {
        "description": "Backtest del motor: cuanto recomienda, cuanto acierta y con que ROI.\n\nEs la validacion de §9 del comparativo y **no se muestra al cliente**.\n\nLa honestidad de la medicion depende de una sola cosa: que los historiales no\nsean los que se usaron para calibrar los pesos. Las sesiones reales lo\ncumplen por construccion (los pesos se fijaron sobre simulaciones, antes de\nque existieran). Las ruedas simuladas usan una semilla propia y sirven de\nlinea base: ahi el ROI TIENE que quedarse en la ventaja de la casa, y si no\nlo hace es que algo esta mal medido.",
        "operationId": "recommendation_backtest_admin_recommendations_backtest_get",
        "parameters": [
          {
            "description": "Variante a evaluar. Por defecto, la primera activa.",
            "in": "query",
            "name": "variant_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Variante a evaluar. Por defecto, la primera activa.",
              "title": "Variant Id"
            }
          },
          {
            "description": "`sessions` corre sobre las sesiones reales de los usuarios; `simulated` sobre ruedas justas generadas, como linea base.",
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/BacktestSource",
              "default": "sessions",
              "description": "`sessions` corre sobre las sesiones reales de los usuarios; `simulated` sobre ruedas justas generadas, como linea base."
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "spins",
            "required": false,
            "schema": {
              "default": 150,
              "maximum": 500,
              "minimum": 20,
              "title": "Spins",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "threshold",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100.0,
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Threshold"
            }
          },
          {
            "in": "query",
            "name": "weak_threshold",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100.0,
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Weak Threshold"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Recommendation Backtest",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/games": {
      "get": {
        "operationId": "list_games_games_get",
        "parameters": [
          {
            "description": "Incluye juegos y variantes desactivados. Solo para administradores.",
            "in": "query",
            "name": "include_inactive",
            "required": false,
            "schema": {
              "default": false,
              "description": "Incluye juegos y variantes desactivados. Solo para administradores.",
              "title": "Include Inactive",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GameResponse"
                  },
                  "title": "Response List Games Games Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Games",
        "tags": [
          "games"
        ]
      }
    },
...
    "/games/{game_id}/variants": {
      "get": {
        "operationId": "list_variants_games__game_id__variants_get",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Game Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GameVariantResponse"
                  },
                  "title": "Response List Variants Games  Game Id  Variants Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Variants",
        "tags": [
          "games"
        ]
      }
    },
...
  }
}